requirements.txt
torch
transformers
accelerate
bitsandbytes
datasets
numpy
scipy
tqdm
pyyaml
ipykernel
protobuf
sentencepiece
